a ton of comments, and other

+++ src/com/discordJava/classes/Snowflake.java
@@ -3,16 +3,16 @@ package com.discordJava.classes;
 import java.sql.Timestamp;
 
 public class Snowflake {
-    public long time;
-    public byte internalWorkerID;
-    public byte internalProcessID;
-    public short processIncrement;
+    public Long time;
+    public Byte internalWorkerID;
+    public Byte internalProcessID;
+    public Short processIncrement;
 
     Snowflake(Long snowflake) {
         this.time = snowflake >> 22;
-        this.internalWorkerID = (byte) ((snowflake >> 17) & 0b11111);
-        this.internalProcessID = (byte) ((snowflake >> 12) & 0b11111);
-        this.processIncrement = (short) (snowflake & 0b111111111111);
+        this.internalWorkerID = ((Long) ((snowflake >> 17) & 0b11111)).byteValue();
+        this.internalProcessID = ((Long) ((snowflake >> 12) & 0b11111)).byteValue();
+        this.processIncrement = ((Long) (snowflake & 0b111111111111)).shortValue();
     }
 
     private Timestamp getDate() {
